refactor(desktop): group services into main/services
