+++ apps/desktop/src/main/ipc/database/backup.ts
@@ -4,7 +4,7 @@ import { IPC_CHANNELS, type DbExportResult, type DbImportResult } from '@shirana
 import { initializeDatabase, closeDatabase } from '@shiranami/database/client';
 import { logger } from '../../app/logger';
 import { handle } from '../with-ipc-handler';
-import { exportDatabase, importDatabase } from '../../db-backup';
+import { exportDatabase, importDatabase } from '../../services/db-backup';
 import { dbBackupExportArgs, dbBackupImportArgs } from '../schemas/db-backup';
 
 const B = IPC_CHANNELS.db.backup;
