+++ apps/desktop/src/main/services/loudness-service.ts
@@ -10,8 +10,8 @@
 
 import * as fs from 'fs';
 import { execFile } from 'child_process';
-import { getFFmpegPath, isFFmpegInstalled } from './ffmpeg-manager';
-import { logger } from './app/logger';
+import { getFFmpegPath, isFFmpegInstalled } from '../ffmpeg-manager';
+import { logger } from '../app/logger';
 
 /** Max time a single loudnorm pass may run before being abandoned (ms). */
 const ANALYZE_TIMEOUT_MS = 120000;
