+++ apps/desktop/src/main/services/metadata-writer.ts
@@ -1,9 +1,9 @@
 import * as path from 'path';
 import * as fs from 'fs';
 import { execFile } from 'child_process';
-import { getFFmpegPath, isFFmpegInstalled } from './ffmpeg-manager';
-import { saveAlbumArt } from './art-protocol';
-import { logger } from './app/logger';
+import { getFFmpegPath, isFFmpegInstalled } from '../ffmpeg-manager';
+import { saveAlbumArt } from '../art-protocol';
+import { logger } from '../app/logger';
 
 export interface WriteMetadataOptions {
   title?: string;
