+++ apps/desktop/src/main/services/weather-service.test.ts
